requirements.txt
dnspython~=2.3.0
paramiko~=3.2.0
pymongo~=4.4.0
python-ldap~=3.4.3